service methods tested, without wrapper methods
