+++ src/test/java/com/greenfox/zerdaReader/service/UpdateServiceTest.java
@@ -6,6 +6,7 @@ package com.greenfox.zerdaReader.service;
  */
 
 import com.greenfox.zerdaReader.ZerdaReaderApplication;
+import com.greenfox.zerdaReader.domain.Feed;
 import com.greenfox.zerdaReader.repository.FeedItemRepository;
 import com.greenfox.zerdaReader.repository.FeedRepository;
 import com.greenfox.zerdaReader.repository.FeedsForUsersRepository;
@@ -21,6 +22,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
+import java.util.Date;
 
 
 @RunWith(SpringJUnit4ClassRunner.class)
@@ -55,7 +57,6 @@ public class UpdateServiceTest {
         Assert.assertEquals(dateTime,feedRepository.findOneByRssPath("file:///D:\\Zerda-reader-Olivin\\src\\test\\resources\\indexrss.xml").getPubDate());
     }
 
-
     @Test
     @Sql({"/clear-tables.sql"})
     public void updateNeededTest() throws Exception {
@@ -86,6 +87,30 @@ public class UpdateServiceTest {
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm");
         LocalDateTime dateTime = LocalDateTime.parse(str, formatter);
         Assert.assertEquals(dateTime,updateService.convertDate(tempSyndFeedStorage.getSyndFeed().getPublishedDate()));
+
+    }
+
+    @Test
+    @Sql({"/clear-tables.sql", "/PopulateTables.sql"})
+    public void TestUpdateNeeded() throws Exception {
+        Feed feed = feedRepository.findOne(2L);
+        LocalDateTime feedDateinDb = feed.getPubDate();
+        TempSyndFeedStorage tempSyndFeedStorage = new TempSyndFeedStorage("file:///D:\\Zerda-reader-Olivin\\src\\test\\resources\\index.xml");
+        Date dateOfXml = tempSyndFeedStorage.getSyndFeed().getPublishedDate();
+        Assert.assertTrue(updateService.isUpdateNeeded(feed,tempSyndFeedStorage.getSyndFeed()));
+    }
+
+    @Test
+    @Sql({"/clear-tables.sql","/PopulateTables2.sql"})
+    public void setPubdateTest() throws Exception {
+        Feed feed = feedRepository.findOne(2L);
+        String rssPath = feed.getRssPath();
+        TempSyndFeedStorage storage = new TempSyndFeedStorage("file:///D:\\Zerda-reader-Olivin\\src\\test\\resources\\index.xml");
+        String str = "2017-02-06 10:23";
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm");
+        LocalDateTime dateTime = LocalDateTime.parse(str, formatter);
+        feed.setPubDate(updateService.convertDate(storage.getSyndFeed().getPublishedDate()));
+        Assert.assertEquals(dateTime,feed.getPubDate());
 
     }
